Integration Tests Spring cleanup
