+++ integration/build_gcb_test.go
@@ -1,0 +1,34 @@
+/*
+Copyright 2019 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package integration
+
+import (
+	"testing"
+
+	"github.com/GoogleContainerTools/skaffold/integration/skaffold"
+)
+
+func TestBuildGCBWithDefaultRepo(t *testing.T) {
+	if testing.Short() || !RunOnGCP() {
+		t.Skip("skipping GCP integration test")
+	}
+
+	// The GCB project (k8s-skaffold) has to be deduced from artifact's image name
+	// after the default repo is applied.
+	// If it's not properly resolved, the build will fail.
+	skaffold.Build("-d", "gcr.io/k8s-skaffold").InDir("testdata/gcb-default-repo").RunOrFail(t)
+}
